Back out Mac port typecasts for compatibility reasons

+++ include/id3/io_strings.h
@@ -63,7 +63,7 @@ namespace dami
       }
       virtual size_type readChars(char_type buf[], size_type len)
       {
-        size_type size = min((unsigned long)len, _string.size() - _cur);
+        size_type size = min(len, _string.size() - _cur);
         _string.copy(reinterpret_cast<String::value_type *>(buf), size, _cur);
         _cur += size;
         return size;
@@ -100,7 +100,7 @@ namespace dami
 
       virtual size_type skipChars(size_type len)
       {
-        size_type size = min((unsigned long)len, _string.size() - _cur);
+        size_type size = min(len, _string.size() - _cur);
         _cur += size;
         return size;
       }
@@ -133,7 +133,7 @@ namespace dami
       }
       virtual size_type readChars(char_type buf[], size_type len)
       {
-        size_type size = min((unsigned long)len, _string.size() - _cur);
+        size_type size = min(len, _string.size() - _cur);
         _string.copy(reinterpret_cast<BString::value_type *>(buf), size, _cur);
         _cur += size;
         return size;
@@ -170,7 +170,7 @@ namespace dami
 
       virtual size_type skipChars(size_type len)
       {
-        size_type size = min((unsigned long)len, _string.size() - _cur);
+        size_type size = min(len, _string.size() - _cur);
         _cur += size;
         return size;
       }
